+++ src/openhuman/agent/mod.rs
@@ -1,5 +1,4 @@
 pub mod agents;
-pub mod classifier;
 pub mod context_pipeline;
 pub mod dispatcher;
 pub mod error;
@@ -10,7 +9,6 @@ pub mod memory_loader;
 pub mod multimodal;
 pub mod prompt;
 mod schemas;
-pub mod traits;
 pub use schemas::{
     all_controller_schemas as all_agent_controller_schemas,
     all_registered_controllers as all_agent_registered_controllers,
